+++ loading-button-android/src/main/java/br/com/simplepass/loading_button_lib/Extensions.kt
@@ -1,9 +1,22 @@
 package br.com.simplepass.loading_button_lib
 
 import android.animation.Animator
+import android.view.View
 
 internal fun Animator.disposeAnimator() {
     end()
     removeAllListeners()
     cancel()
+}
+
+internal fun View.updateWidth(width: Int) {
+    val layoutParams = this.layoutParams
+    layoutParams.width = width
+    this.layoutParams = layoutParams
+}
+
+internal fun View.updateHeight(height: Int) {
+    val layoutParams = this.layoutParams
+    layoutParams.height = height
+    this.layoutParams = layoutParams
 }
